requirements.txt
Flask==3.0.0
Jinja2==3.1.2
requests==2.31.0
gunicorn==21.2.0
boto3==1.34.2
botocore==1.34.2
python-dotenv==1.0.1
